Allow builds for armv7l/armhf.
electron-builder currently for snap builds is doing crazy unneeded and 
weird stuff for arm64 (using multipass) which I will bug them about to 
build as normal. For now we can use this as a start.

I know armv7l and armhf is 32-bit but I really don't care, it's for 
testing purposes and soon I'll probably remove it.